+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - `Arg.setSetting` is renamed to `Arg.addProperty`.
 - `Arg.unsetSetting` is renamed to `Arg.removeProperty`.
 - `Arg.isSettingSet` is renamed to `Arg.hasProperty`.
+- `Command.takesSingleValue` and `Command.takesNValues` are removed, use new `Arg.positional` instead.
 - `Command.countArgs` is renamed to `Command.countPositionalArgs`.
 - `Command.setSetting` is renamed to `Command.addProperty`.
 - `Command.unsetSetting` is renamed to `Command.removeProperty`.
@@ -48,6 +49,38 @@
 
 ## What's New
 - Added new `Arg.multiArgumentsOptionWithValidValues` API
+- Added new `Arg.positional` API for creating a new positional argument and with this changes it's no
+longer required to set `.takes_positional_arg` property to root `Command`.
+
+    Before
+    ```zig
+    var root = app.rootCommand();
+
+    try root.takesSingleValue("ONE");
+    try root.takesSingleValue("TWO");
+    try root.takesSingleValue("THREE");
+    root.setSetting(.takes_positional_arg);
+    ```
+
+    After
+    ```zig
+    const Arg = yazap.Arg;
+
+    var root = app.rootCommand();
+
+    // Arg.positional(name: []const u8, description: ?[]const u8, index: ?usize)
+    
+    // Order dependent
+    try root.addArg(Arg.positional("ONE", null, null));
+    try root.addArg(Arg.positional("TWO", null, null));
+    try root.addArg(Arg.positional("THREE", null, null));
+
+    // Equivalent but order independent
+    try root.addArg(Arg.positional("THREE", null, 3));
+    try root.addArg(Arg.positional("TWO", null, 2));
+    try root.addArg(Arg.positional("ONE", null, 1));
+
+    ```
 
 ## Internal Changes
 - `enable_help` property is removed and no longer needed to set for a command